requirements.txt
scapy>=2.5.0
flask>=3.0.0
flask-socketio>=5.3.0
pyyaml>=6.0
requests>=2.31.0
dnspython>=2.6.0
psutil>=5.9.0
eventlet>=0.36.0
colorama>=0.4.6
watchdog>=4.0.0
